+++ src/main/java/edu/ie3/datamodel/io/factory/result/ResultEntityFactory.java
@@ -22,6 +22,7 @@ public abstract class ResultEntityFactory<T extends ResultEntity>
     extends EntityFactory<T, EntityData> {
 
   protected static final String TIME = "time";
+  protected static final String INPUT_MODEL = "inputModel";
 
   protected final TimeUtil timeUtil;