+++ src/main/java/edu/ie3/datamodel/io/sink/SqlSink.java
@@ -5,10 +5,12 @@
 */
 package edu.ie3.datamodel.io.sink;
 
-import static edu.ie3.datamodel.io.SqlUtils.*;
+import static edu.ie3.datamodel.io.SqlUtils.quote;
 import static java.util.stream.Collectors.groupingBy;
 
-import edu.ie3.datamodel.exceptions.*;
+import edu.ie3.datamodel.exceptions.EntityProcessorException;
+import edu.ie3.datamodel.exceptions.ExtractorException;
+import edu.ie3.datamodel.exceptions.ProcessorProviderException;
 import edu.ie3.datamodel.io.DbGridMetadata;
 import edu.ie3.datamodel.io.connectors.SqlConnector;
 import edu.ie3.datamodel.io.extractor.Extractor;
@@ -18,17 +20,22 @@ import edu.ie3.datamodel.io.processor.EntityProcessor;
 import edu.ie3.datamodel.io.processor.ProcessorProvider;
 import edu.ie3.datamodel.io.processor.timeseries.TimeSeriesProcessorKey;
 import edu.ie3.datamodel.models.Entity;
-import edu.ie3.datamodel.models.input.*;
-import edu.ie3.datamodel.models.input.connector.*;
+import edu.ie3.datamodel.models.input.AssetTypeInput;
+import edu.ie3.datamodel.models.input.InputEntity;
+import edu.ie3.datamodel.models.input.NodeInput;
+import edu.ie3.datamodel.models.input.OperatorInput;
+import edu.ie3.datamodel.models.input.connector.ConnectorInput;
 import edu.ie3.datamodel.models.input.container.JointGridContainer;
 import edu.ie3.datamodel.models.input.graphics.GraphicInput;
-import edu.ie3.datamodel.models.input.system.*;
+import edu.ie3.datamodel.models.input.system.SystemParticipantInput;
 import edu.ie3.datamodel.models.input.thermal.ThermalBusInput;
 import edu.ie3.datamodel.models.input.thermal.ThermalUnitInput;
 import edu.ie3.datamodel.models.result.ResultEntity;
 import edu.ie3.datamodel.models.timeseries.TimeSeries;
 import edu.ie3.datamodel.models.timeseries.TimeSeriesEntry;
+import edu.ie3.datamodel.models.timeseries.repetitive.LoadProfileTimeSeries;
 import edu.ie3.datamodel.models.value.Value;
+import edu.ie3.datamodel.utils.TriFunction;
 import edu.ie3.util.StringUtils;
 import java.sql.SQLException;
 import java.util.*;
@@ -47,6 +54,7 @@ public class SqlSink {
   private final String schemaName;
 
   private static final String TIME_SERIES = "time_series";
+  private static final String LOAD_PROFILE = "load_profile";
 
   public SqlSink(
       String schemaName, DatabaseNamingStrategy databaseNamingStrategy, SqlConnector connector)
@@ -109,19 +117,19 @@ public class SqlSink {
   }
 
   /**
-   * Persist an entity. By default this method takes care of the extraction process of nested
+   * Persist an entity. By default, this method takes care of the extraction process of nested
    * entities (if any)
    *
    * @param entity the entity that should be persisted
    * @param identifier identifier of the grid
-   * @throws SQLException
+   * @throws SQLException if an error occurred
    */
   public <C extends Entity> void persist(C entity, DbGridMetadata identifier) throws SQLException {
     if (entity instanceof InputEntity inputEntity) {
       persistIncludeNested(inputEntity, identifier);
     } else if (entity instanceof ResultEntity resultEntity) {
       insert(resultEntity, identifier);
-    } else if (entity instanceof TimeSeries<?, ?> timeSeries) {
+    } else if (entity instanceof TimeSeries<?, ?, ?> timeSeries) {
       persistTimeSeries(timeSeries, identifier);
     } else {
       log.error(
@@ -178,8 +186,9 @@ public class SqlSink {
    * Persist a list of entities with same types. To minimize the number of queries, the entities
    * will be grouped by their class.
    */
-  private <C extends Entity, E extends TimeSeriesEntry<V>, V extends Value> void persistList(
-      List<C> entities, Class<C> cls, DbGridMetadata identifier) throws SQLException {
+  private <C extends Entity, E extends TimeSeriesEntry<V>, V extends Value, R extends Value>
+      void persistList(List<C> entities, Class<C> cls, DbGridMetadata identifier)
+          throws SQLException {
     // Check if there are only elements of the same class
     Class<?> firstClass = entities.get(0).getClass();
     boolean allSameClass = entities.stream().allMatch(e -> e.getClass() == firstClass);
@@ -190,7 +199,7 @@ public class SqlSink {
       } else if (ResultEntity.class.isAssignableFrom(cls)) {
         insertListIgnoreNested(entities, cls, identifier, false);
       } else if (TimeSeries.class.isAssignableFrom(cls)) {
-        entities.forEach(ts -> persistTimeSeries((TimeSeries<E, V>) ts, identifier));
+        entities.forEach(ts -> persistTimeSeries((TimeSeries<E, V, R>) ts, identifier));
       } else {
         log.error("I don't know how to handle an entity of class {}", cls.getSimpleName());
       }
@@ -222,8 +231,8 @@ public class SqlSink {
   }
 
   /** Persist one time series. */
-  protected <E extends TimeSeriesEntry<V>, V extends Value> void persistTimeSeries(
-      TimeSeries<E, V> timeSeries, DbGridMetadata identifier) {
+  protected <E extends TimeSeriesEntry<V>, V extends Value, R extends Value> void persistTimeSeries(
+      TimeSeries<E, V, R> timeSeries, DbGridMetadata identifier) {
     try {
       TimeSeriesProcessorKey key = new TimeSeriesProcessorKey(timeSeries);
       String[] headerElements = processorProvider.getHeaderElements(key);
@@ -234,17 +243,30 @@ public class SqlSink {
     }
   }
 
-  private <E extends TimeSeriesEntry<V>, V extends Value> void persistTimeSeries(
-      TimeSeries<E, V> timeSeries, String[] headerElements, DbGridMetadata identifier)
+  private <E extends TimeSeriesEntry<V>, V extends Value, R extends Value> void persistTimeSeries(
+      TimeSeries<E, V, R> timeSeries, String[] headerElements, DbGridMetadata identifier)
       throws ProcessorProviderException {
     try {
+
+      TriFunction<String, String, String[], String> queryBuilder;
+      String timeSeriesIdentifier;
+
+      if (timeSeries instanceof LoadProfileTimeSeries<?> lpts) {
+        timeSeriesIdentifier = lpts.getLoadProfile().getKey();
+        queryBuilder = this::basicInsertQueryValuesLPTS;
+      } else {
+        timeSeriesIdentifier = timeSeries.getUuid().toString();
+        queryBuilder = this::basicInsertQueryValuesITS;
+      }
+
       String query =
-          basicInsertQueryValuesITS(
+          queryBuilder.apply(
               schemaName,
               databaseNamingStrategy.getEntityName(timeSeries).orElseThrow(),
               headerElements);
       Set<LinkedHashMap<String, String>> entityFieldData =
           processorProvider.handleTimeSeries(timeSeries);
+
       query =
           query
               + entityFieldData.stream()
@@ -254,7 +276,7 @@ public class SqlSink {
                               sqlEntityFieldData(data),
                               headerElements,
                               identifier,
-                              timeSeries.getUuid().toString()))
+                              timeSeriesIdentifier))
                   .collect(Collectors.joining(",\n", "", ";"));
       executeQueryToPersist(query);
     } catch (ProcessorProviderException e) {
@@ -352,13 +374,13 @@ public class SqlSink {
       Map<String, String> entityFieldData,
       String[] headerElements,
       DbGridMetadata identifier,
-      String tsUuid) {
+      String timeSeriesIdentifier) {
     return writeOneLine(
         Stream.concat(
             Stream.concat(
                 Arrays.stream(headerElements).map(entityFieldData::get),
                 identifier.getStreamForQuery()),
-            Stream.of(quote(tsUuid, "'"))));
+            Stream.of(quote(timeSeriesIdentifier, "'"))));
   }
 
   private LinkedHashMap<String, String> sqlEntityFieldData(
@@ -378,10 +400,8 @@ public class SqlSink {
   private String basicInsertQueryValuesGrid(
       String schemaName, String tableName, String[] headerElements) {
     String[] addParams = {DbGridMetadata.GRID_UUID_COLUMN};
-    return basicInsertQuery(schemaName, tableName)
-        + " "
-        + writeOneLine(StringUtils.camelCaseToSnakeCase(headerElements), addParams)
-        + "\nVALUES\n";
+    return basicInsertQueryWith(
+        schemaName, tableName, StringUtils.camelCaseToSnakeCase(headerElements), addParams);
   }
 
   /**
@@ -391,6 +411,21 @@ public class SqlSink {
   private String basicInsertQueryValuesITS(
       String schemaName, String tableName, String[] headerElements) {
     String[] addParams = {DbGridMetadata.GRID_UUID_COLUMN, TIME_SERIES};
+    return basicInsertQueryWith(
+        schemaName, tableName, StringUtils.camelCaseToSnakeCase(headerElements), addParams);
+  }
+
+  /** Provides the insert, column names, grid identifier, and the VALUES statement for a query. */
+  private String basicInsertQueryValuesLPTS(
+      String schemaName, String tableName, String[] headerElements) {
+    String[] addParams = {DbGridMetadata.GRID_UUID_COLUMN, LOAD_PROFILE};
+    return basicInsertQueryWith(
+        schemaName, tableName, StringUtils.camelCaseToSnakeCase(headerElements), addParams);
+  }
+
+  /** Provides the insert, column names, grid identifier, and the VALUES statement for a query */
+  private String basicInsertQueryWith(
+      String schemaName, String tableName, String[] headerElements, String[] addParams) {
     return basicInsertQuery(schemaName, tableName)
         + " "
         + writeOneLine(StringUtils.camelCaseToSnakeCase(headerElements), addParams)